Remove all references to gax
